Improve "bootstrap idempotent" test
Since this test is run post-install, running it twice is unnecessary.

Issue: GH-881
