+++ tests/post/features/bootstrap.feature
@@ -1,6 +1,6 @@
 @post @ci @local
 Feature: Bootstrap
     Scenario: Re-run bootstrap
-        Given bootstrap was run once
+        Given the Kubernetes API is available
         When we run bootstrap a second time
         Then the Kubernetes API is available
